64-bit support

+++ lib/inc/rwt_init.h
@@ -12,8 +12,8 @@
     #include "matrix.h"
   #endif
   typedef struct {
-    int nrows;        /*!< The number of rows in the input matrix. Output matrix will match.  */
-    int ncols;        /*!< The number of columns in the input matrix. Output matrix will match. */
+    size_t nrows;     /*!< The number of rows in the input matrix. Output matrix will match.  */
+    size_t ncols;     /*!< The number of columns in the input matrix. Output matrix will match. */
     int levels;       /*!< L, the number of levels for the transform. */
     int lh;           /*!< Length of h / the number of scaling coefficients */
     double *scalings; /*!< Wavelet scaling coefficients */
@@ -28,8 +28,8 @@ extern "C" {
 #ifdef MATLAB_MEX_FILE
   rwt_init_params rwt_matlab_init(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[], transform_t dwtType);
 #else
-  int rwt_find_levels(int m, int n);
-  int rwt_check_levels(int levels, int rows, int cols);
+  int rwt_find_levels(size_t m, size_t n);
+  int rwt_check_levels(int levels, size_t rows, size_t cols);
 #endif
 
 #ifdef __cplusplus
